将各种ID字段改成int64

+++ pkg/distlock/reqbuilder/metadata_bucket.go
@@ -13,7 +13,7 @@ func (b *MetadataLockReqBuilder) Bucket() *MetadataBucketLockReqBuilder {
 	return &MetadataBucketLockReqBuilder{MetadataLockReqBuilder: b}
 }
 
-func (b *MetadataBucketLockReqBuilder) ReadOne(bucketID int) *MetadataBucketLockReqBuilder {
+func (b *MetadataBucketLockReqBuilder) ReadOne(bucketID int64) *MetadataBucketLockReqBuilder {
 	b.locks = append(b.locks, distlock.Lock{
 		Path:   b.makePath("Bucket"),
 		Name:   lockprovider.METADATA_ELEMENT_READ_LOCK,
@@ -21,7 +21,7 @@ func (b *MetadataBucketLockReqBuilder) ReadOne(bucketID int) *MetadataBucketLock
 	})
 	return b
 }
-func (b *MetadataBucketLockReqBuilder) WriteOne(bucketID int) *MetadataBucketLockReqBuilder {
+func (b *MetadataBucketLockReqBuilder) WriteOne(bucketID int64) *MetadataBucketLockReqBuilder {
 	b.locks = append(b.locks, distlock.Lock{
 		Path:   b.makePath("Bucket"),
 		Name:   lockprovider.METADATA_ELEMENT_WRITE_LOCK,
@@ -29,7 +29,7 @@ func (b *MetadataBucketLockReqBuilder) WriteOne(bucketID int) *MetadataBucketLoc
 	})
 	return b
 }
-func (b *MetadataBucketLockReqBuilder) CreateOne(userID int, bucketName string) *MetadataBucketLockReqBuilder {
+func (b *MetadataBucketLockReqBuilder) CreateOne(userID int64, bucketName string) *MetadataBucketLockReqBuilder {
 	b.locks = append(b.locks, distlock.Lock{
 		Path:   b.makePath("Bucket"),
 		Name:   lockprovider.METADATA_ELEMENT_CREATE_LOCK,
