+++ pkg/distlock/reqbuilder/metadata_user_bucket.go
@@ -13,7 +13,7 @@ func (b *MetadataLockReqBuilder) UserBucket() *MetadataUserBucketLockReqBuilder 
 	return &MetadataUserBucketLockReqBuilder{MetadataLockReqBuilder: b}
 }
 
-func (b *MetadataUserBucketLockReqBuilder) ReadOne(userID int, bucketID int) *MetadataUserBucketLockReqBuilder {
+func (b *MetadataUserBucketLockReqBuilder) ReadOne(userID int64, bucketID int64) *MetadataUserBucketLockReqBuilder {
 	b.locks = append(b.locks, distlock.Lock{
 		Path:   b.makePath("UserBucket"),
 		Name:   lockprovider.METADATA_ELEMENT_READ_LOCK,
@@ -21,7 +21,7 @@ func (b *MetadataUserBucketLockReqBuilder) ReadOne(userID int, bucketID int) *Me
 	})
 	return b
 }
-func (b *MetadataUserBucketLockReqBuilder) WriteOne(userID int, bucketID int) *MetadataUserBucketLockReqBuilder {
+func (b *MetadataUserBucketLockReqBuilder) WriteOne(userID int64, bucketID int64) *MetadataUserBucketLockReqBuilder {
 	b.locks = append(b.locks, distlock.Lock{
 		Path:   b.makePath("UserBucket"),
 		Name:   lockprovider.METADATA_ELEMENT_WRITE_LOCK,
@@ -29,7 +29,7 @@ func (b *MetadataUserBucketLockReqBuilder) WriteOne(userID int, bucketID int) *M
 	})
 	return b
 }
-func (b *MetadataUserBucketLockReqBuilder) CreateOne(userID int, bucketID int) *MetadataUserBucketLockReqBuilder {
+func (b *MetadataUserBucketLockReqBuilder) CreateOne(userID int64, bucketID int64) *MetadataUserBucketLockReqBuilder {
 	b.locks = append(b.locks, distlock.Lock{
 		Path:   b.makePath("UserBucket"),
 		Name:   lockprovider.METADATA_ELEMENT_CREATE_LOCK,
